+++ test/registered/ascend/performance/test_npu_qwen3_5_27b_w8a8_1p_in128k_90prefix_50ms.py
