+++ test/registered/ascend/performance/test_npu_qwen3_5_27b_w8a8_1p_in16k_out1k_20ms.py
@@ -10,11 +10,11 @@ from sglang.test.ci.ci_register import register_npu_ci
 
 register_npu_ci(
     est_time=1800,
-    suite="nightly-4-npu-a3",
+    suite="nightly-2-npu-a3",
     nightly=True,
 )
 
-QWEN3_5_27B_256K_1K_ENVS = {
+QWEN3_5_27B_16K_1K_LOW_ENVS = {
     "PYTORCH_NPU_ALLOC_CONF": "expandable_segments:True",
     "STREAMS_PER_DEVICE": "32",
     "HCCL_SOCKET_IFNAME": "lo",
@@ -25,13 +25,14 @@ QWEN3_5_27B_256K_1K_ENVS = {
     "SGLANG_ENABLE_OVERLAP_PLAN_STREAM": "0",
     "SGLANG_SCHEDULER_DECREASE_PREFILL_IDLE": "1",
     "SGLANG_PREFILL_DELAYER_MAX_DELAY_PASSES": "100",
+    "SGLANG_NPU_PROFILING": "1",
 }
 
-QWEN3_5_27B_256K_1K_OTHER_ARGS = [
+QWEN3_5_27B_16K_1K_LOW_OTHER_ARGS = [
     "--model-path",
     QWEN3_5_27B_W8A8_HOME_MODEL_PATH,
     "--tp-size",
-    4,
+    2,
     "--nnodes",
     1,
     "--node-rank",
@@ -43,17 +44,21 @@ QWEN3_5_27B_256K_1K_OTHER_ARGS = [
     "--chunked-prefill-size",
     -1,
     "--max-prefill-tokens",
-    256000,
+    65000,
     "--disable-radix-cache",
     "--trust-remote-code",
     "--max-running-requests",
-    1,
+    32,
     "--max-mamba-cache-size",
-    100,
+    32,
     "--mem-fraction-static",
-    0.6,
+    0.85,
     "--cuda-graph-bs",
-    1,
+    2,
+    3,
+    4,
+    5,
+    6,
     "--enable-multimodal",
     "--quantization",
     "modelslim",
@@ -64,7 +69,7 @@ QWEN3_5_27B_256K_1K_OTHER_ARGS = [
     "--mamba-ssm-dtype",
     "bfloat16",
     "--max-total-tokens",
-    850000,
+    310000,
     "--speculative-algorithm",
     "NEXTN",
     "--speculative-num-steps",
@@ -76,25 +81,25 @@ QWEN3_5_27B_256K_1K_OTHER_ARGS = [
 ]
 
 
-class TestNPUQwen3_5_27B_2P_In256k_Out1k(TestAscendPerformanceTestCaseBase):
-    """Test NPU performance for Qwen3.5-27B-W8A8 2p in256k out1k"""
+class TestNPUQwen3_5_27B_1P_In16k_Out1k_Low(TestAscendPerformanceTestCaseBase):
+    """Test NPU performance for Qwen3.5-27B-W8A8 1p in16k out1k low latency"""
 
     benchmark_tool = BENCHMARK_TOOL_DEFAULT
     aisbench_dataset_type = AISBENCHMARK_DATASET_DEFAULT
     model = QWEN3_5_27B_W8A8_HOME_MODEL_PATH
-    other_args = QWEN3_5_27B_256K_1K_OTHER_ARGS
-    envs = QWEN3_5_27B_256K_1K_ENVS
+    other_args = QWEN3_5_27B_16K_1K_LOW_OTHER_ARGS
+    envs = QWEN3_5_27B_16K_1K_LOW_ENVS
     dataset_name = "random"
-    max_concurrency = 1
-    num_prompts = 1
-    input_len = 250000
+    max_concurrency = 4
+    num_prompts = 16
+    input_len = 16384
     output_len = 1024
     random_range_ratio = 1
-    tpot = 50
-    output_token_throughput = 50
+    tpot = 20
+    output_token_throughput = 200
 
-    def test_npu_qwen3_5_27b_2p_in256k_out1k(self):
-        """Run NPU performance test for Qwen3.5-27B-W8A8 in256k out1k"""
+    def test_npu_qwen3_5_27b_1p_in16k_out1k_low(self):
+        """Run NPU performance test for Qwen3.5-27B-W8A8 in16k out1k low latency"""
         self.run_throughput()
