+++ test/registered/ascend/performance/test_npu_qwen3_5_27b_w8a8_1p_in3k5_out1k5_50ms.py
