+++ test/registered/ascend/performance/test_npu_qwen3_5_27b_w8a8_1p_in64k_90prefix_20ms.py
@@ -14,7 +14,7 @@ register_npu_ci(
     nightly=True,
 )
 
-QWEN3_5_27B_16K_1K_LOW_ENVS = {
+QWEN3_5_27B_64K_90_PREFIX_LOW_ENVS = {
     "PYTORCH_NPU_ALLOC_CONF": "expandable_segments:True",
     "STREAMS_PER_DEVICE": "32",
     "HCCL_SOCKET_IFNAME": "lo",
@@ -23,12 +23,10 @@ QWEN3_5_27B_16K_1K_LOW_ENVS = {
     "SGLANG_SET_CPU_AFFINITY": "1",
     "SGLANG_ENABLE_SPEC_V2": "1",
     "SGLANG_ENABLE_OVERLAP_PLAN_STREAM": "0",
-    "SGLANG_SCHEDULER_DECREASE_PREFILL_IDLE": "1",
-    "SGLANG_PREFILL_DELAYER_MAX_DELAY_PASSES": "100",
     "SGLANG_NPU_PROFILING": "1",
 }
 
-QWEN3_5_27B_16K_1K_LOW_OTHER_ARGS = [
+QWEN3_5_27B_64K_90_PREFIX_LOW_OTHER_ARGS = [
     "--model-path",
     QWEN3_5_27B_W8A8_HOME_MODEL_PATH,
     "--tp-size",
@@ -44,21 +42,20 @@ QWEN3_5_27B_16K_1K_LOW_OTHER_ARGS = [
     "--chunked-prefill-size",
     -1,
     "--max-prefill-tokens",
-    65000,
-    "--disable-radix-cache",
+    68000,
+    "--mamba-scheduler-strategy",
+    "extra_buffer",
     "--trust-remote-code",
     "--max-running-requests",
     32,
     "--max-mamba-cache-size",
-    32,
+    18,
     "--mem-fraction-static",
-    0.85,
+    0.5,
     "--cuda-graph-bs",
     2,
     3,
     4,
-    5,
-    6,
     "--enable-multimodal",
     "--quantization",
     "modelslim",
@@ -68,8 +65,6 @@ QWEN3_5_27B_16K_1K_LOW_OTHER_ARGS = [
     "bfloat16",
     "--mamba-ssm-dtype",
     "bfloat16",
-    "--max-total-tokens",
-    310000,
     "--speculative-algorithm",
     "NEXTN",
     "--speculative-num-steps",
@@ -81,25 +76,25 @@ QWEN3_5_27B_16K_1K_LOW_OTHER_ARGS = [
 ]
 
 
-class TestNPUQwen3_5_27B_1P_In16k_Out1k_Low(TestAscendPerformanceTestCaseBase):
-    """Test NPU performance for Qwen3.5-27B-W8A8 1p in16k out1k low latency"""
+class TestNPUQwen3_5_27B_1P_In64k_90Prefix_Low(TestAscendPerformanceTestCaseBase):
+    """Test NPU performance for Qwen3.5-27B-W8A8 1p in64k 90% prefix low latency"""
 
     benchmark_tool = BENCHMARK_TOOL_DEFAULT
     aisbench_dataset_type = AISBENCHMARK_DATASET_DEFAULT
     model = QWEN3_5_27B_W8A8_HOME_MODEL_PATH
-    other_args = QWEN3_5_27B_16K_1K_LOW_OTHER_ARGS
-    envs = QWEN3_5_27B_16K_1K_LOW_ENVS
+    other_args = QWEN3_5_27B_64K_90_PREFIX_LOW_OTHER_ARGS
+    envs = QWEN3_5_27B_64K_90_PREFIX_LOW_ENVS
     dataset_name = "random"
-    max_concurrency = 4
-    num_prompts = 16
-    input_len = 16384
+    max_concurrency = 1
+    num_prompts = 4
+    input_len = 65536
     output_len = 1024
-    random_range_ratio = 1
-    tpot = 50
-    output_token_throughput = 200
+    random_range_ratio = 0.1
+    tpot = 20
+    output_token_throughput = 50
 
-    def test_npu_qwen3_5_27b_1p_in16k_out1k_low(self):
-        """Run NPU performance test for Qwen3.5-27B-W8A8 in16k out1k low latency"""
+    def test_npu_qwen3_5_27b_1p_in64k_90prefix_low(self):
+        """Run NPU performance test for Qwen3.5-27B-W8A8 in64k 90% prefix low latency"""
         self.run_throughput()
