+++ test/registered/ascend/performance/test_npu_qwen3_5_27b_w8a8_1p_mm_1920x1080_50ms.py
