+++ test/registered/ascend/performance/test_npu_qwen3_5_27b_w8a8_2p_in64k_out1k_50ms.py
